+++ main.py
@@ -2,6 +2,8 @@ from matplotlib.figure import Figure
 import japanize_matplotlib as _
 import numpy as np
 
+from regressor import PolyRegressor
+
 def calculate_score(y, y_pred, score_eps):
     norm_y = np.sum(np.abs(y))
     norm_diff = np.sum(np.abs(y - y_pred))
@@ -32,21 +34,6 @@ def save_graph(
         ax.plot(x_pred, y_pred, label=r'回帰関数 $\hat{f}$')
     ax.legend()
     fig.savefig(filename)
-
-class PolyRegressor:
-    def __init__(self, d):
-        self.d = d
-        self.p = np.arange(d+1)[np.newaxis, :] # 1 x (d+1)
-
-    def fit(self, x_sample, y_sample):
-        X_sample = x_sample[:, np.newaxis] ** self.p
-        XX_inv_sample = np.linalg.inv(X_sample.T @ X_sample)
-        self.a =  XX_inv_sample @ X_sample.T @ y_sample[:, np.newaxis]
-
-    def predict(self, x):
-        X = x[:, np.newaxis] ** self.p
-        y_pred = np.squeeze(X @ self.a)
-        return y_pred
 
 def main():
     # 実験条件
